+++ src/internal/naive_trie/naive_trie_b_f_iter.rs
