+++ src/internal/naive_trie/naive_trie_impl.rs
